+++ package.json
@@ -2,15 +2,14 @@
   "name": "react",
   "version": "1.0.0",
   "description": "React example starter project",
-  "keywords": [
-    "react",
-    "starter"
-  ],
-  "main": "src/index.js",
+  "keywords": ["react", "starter"],
+  "main": "src/index.jsx",
   "dependencies": {
     "@reduxjs/toolkit": "1.6.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-markdown": "6.0.2",
+    "react-redux": "7.2.4",
     "react-scripts": "4.0.0"
   },
   "devDependencies": {
@@ -23,10 +22,5 @@
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "browserslist": [">0.2%", "not dead", "not ie <= 11", "not op_mini all"]
 }
