+++ src/App.css
@@ -1,0 +1,42 @@
+:root {
+  --color-background: #fcfcfc;
+  --color-background-secondary: #f3f3f3;
+  --color-foreground: #ffffff;
+  --color-border: #e6e6e6;
+  --color-text-header: #444444;
+  --color-text-body: #6f6f6f;
+  --color-text-secondary: #717171;
+  --color-text-invert: #ffffff;
+  --color-text-link: #4ec3f5;
+  --color-branding: #3d5af1;
+  --color-branding-transparent: #3d5af140;
+
+  --color-alert: #ff304f;
+  --color-success: #45b81f;
+  --color-on-hover: #0000000f;
+
+  --box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1),
+    0 4px 6px -2px rgba(0, 0, 0, 0.05) !important;
+  --box-shadow-hover: 0 20px 25px -5px rgba(0, 0, 0, 0.1),
+    0 10px 10px -5px rgba(0, 0, 0, 0.04) !important;
+
+  --spacing-0: 4px;
+  --spacing-1: 8px;
+  --spacing-2: 16px;
+  --spacing-3: 24px;
+  --spacing-4: 36px;
+  --spacing-5: 64px;
+
+  --radius: 4px;
+  --radius-2: 8px;
+}
+
+html,
+body {
+  background: var(--color-background);
+  color: var(--color-text-body);
+}
+
+p {
+  margin: 0;
+}
