+++ src/componets/Card.css
@@ -1,0 +1,13 @@
+.card {
+  background: var(--color-foreground);
+  box-shadow: var(--box-shadow);
+  margin-bottom: var(--spacing-4);
+  border-radius: var(--radius);
+  padding: var(--spacing-3);
+  transition: box-shadow 0.1s ease-in;
+  color: var(--color-text-secondary);
+}
+
+.card:hover {
+  box-shadow: var(--box-shadow-hover);
+}
