+++ src/features/Comment/Comment.css
@@ -1,0 +1,27 @@
+.comment {
+  background: var(--color-background);
+  margin: var(--spacing-1) 0;
+  padding: var(--spacing-2);
+  transition: box-shadow 0.1s ease-in;
+  border-radius: var(--border-radius);
+}
+
+.comment:hover {
+  box-shadow: var(--box-shadow);
+}
+
+.comment-metadata {
+  display: flex;
+  margin-bottom: var(--spacing-2);
+  align-items: center;
+}
+
+.comment-created-time {
+  margin-left: auto;
+  font-style: italic;
+}
+
+.comment-author {
+  font-weight: bold;
+  color: var(--color-branding);
+}
