+++ src/features/SubReddit/Subreddit.css
@@ -1,0 +1,46 @@
+.subreddit-card {
+  padding-right: var(--spacing-2);
+}
+
+.subreddit-card h2 {
+  margin-top: 0;
+  color: var(--color-text-header);
+}
+
+.subreddit-card button {
+  background: none;
+  border: none;
+  display: flex;
+  align-items: center;
+  color: var(--color-text-body);
+  font-weight: 600;
+  cursor: pointer;
+  width: 100%;
+  padding: var(--spacing-2);
+  border-radius: var(--radius);
+}
+
+.selected-subreddit button {
+  color: var(--color-branding) !important;
+}
+
+.selected-subreddit {
+  border-left: 5px solid var(--color-branding);
+  background: var(--color-branding-transparent);
+}
+
+.subreddit-card button:hover {
+  background: var(--color-on-hover);
+}
+
+.subreddits-list {
+  list-style: none;
+  padding: 0;
+}
+
+.subreddit-icon {
+  height: 25px;
+  width: 25px;
+  border-radius: 50%;
+  margin-right: var(--spacing-1);
+}
