+++ src/index.css
@@ -1,0 +1,42 @@
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+#root {
+  display: grid;
+  grid-template-columns: 1fr 2fr 1fr;
+  grid-gap: var(--spacing-2);
+}
+
+main {
+  padding-left: var(--spacing-2);
+  grid-column: 1/3;
+}
+
+a {
+  color: var(--color-text-link);
+}
+
+header {
+  grid-column: 1/-1;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  grid-gap: var(--spacing-2);
+  align-content: center;
+}
+
+@media only screen and (max-width: 600px) {
+  #root {
+    grid-template-columns: 1fr;
+  }
+
+  main {
+    grid-column: 1;
+    padding: 1rem;
+  }
+}
